Added space freed by the script

+++ windows-diskspace-cleaning-automation.py
@@ -8,18 +8,28 @@ def ClearTemp():
     PATH = f"C:\\Users\\admin\\AppData\\Local\\Temp"
     files = os.listdir(PATH)
     count = 0
+    size = 0
     for file in files:
         try:
             FILE_PATH = PATH + '\\' + file
             if os.path.isfile(FILE_PATH) and os.path.exists(FILE_PATH):
+                if Exception:
+                    continue
+                size += os.path.getsize(FILE_PATH)
                 os.remove(FILE_PATH)
                 count += 1
             elif os.path.isdir(FILE_PATH) and os.path.exists(FILE_PATH):
+                if Exception:
+                    continue
+                size += os.path.getsize(FILE_PATH)
                 shutil.rmtree(FILE_PATH)
                 count += 1
         except Exception:
             continue
-    print(f"Deleted {count} files")
+    print("*"*15)
+    print(f"Deleted {count} files from Temp")
+    print(f"{round(size / 1048576, 2)} MB files cleared from Temp Folder")
+    print("*"*15)
 ############################################################################################################################################################################################## 
 def check_pc_health():
     if ps.virtual_memory()[2] > 65:
@@ -48,18 +58,21 @@ def CleanDownloads():
     sec_per_day = 86400
     number_of_days = 30
     file_count = 0
+    size = 0
     time_now = time.time()
     try:
         for file in files:
             target_file_path = target_path + '\\' + file
             file_age = os.stat(target_file_path).st_mtime
             if file_age < time_now - (sec_per_day*number_of_days):
+                size += os.path.getsize(target_file_path)
                 os.remove(target_file_path)
                 file_count += 1
     except Exception:
         print("No files older than 30 days were found.")
         pass
     print(f"{file_count} files were deleted and were older than 30 days.")
+    print(f"{round(size / 1048576, 2)} MB files cleared from Downloads Folder")
 ####################################################################################################################################################################################################           
 def main():
     ClearTemp()
